requirements.txt
django
djangorestframework
postgresql
requests
node
django-cors-headers
react
axios
react-router-dom
scikit-learn
pandas
pytorch
